Autonomous Mobile Robotics
Utilizing a particle filter algorithm approach to Simultaneously Localize and Map (SLAM) the robot in simulated virtual environment. This was my final submission for the Autonomous Mobile Robotics, a course for third year masters student in smart tech program. With this code, was able to navigate and localize fastest in the course and got a prize.